fix(build): resolve kotlin type mismatches, add generateEmbedding API, and rename shadowed variable in PlayerViewModel

+++ app/src/main/java/com/streamify/app/data/TrackRepository.kt
@@ -70,7 +70,7 @@ object TrackRepository {
             durationSec = track.durationSec,
             coverArtPath = track.coverArtPath ?: "",
             lyricsPath = track.lyricsPath ?: "",
-            bpm = track.bpm.toDouble(),
+            bpm = track.bpm,
             key = track.key
         )
